added DTO's, handled Exceptions & implemented commute table and returning mock data
