--- Backend/src/main/java/com/GreenCommute/Backend/BackendApplication.java
@@ -1,13 +1,0 @@
-package com.GreenCommute.Backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackendApplication.class, args);
-	}
-
-}
