--- Backend/src/main/java/com/GreenCommute/Backend/Dto/JobsDto.java
@@ -1,20 +1,0 @@
-package com.GreenCommute.Backend.Dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class JobsDto {
-
-    private int jobId;
-
-    private String jobTitle;
-
-    private String jobDescription;
-
-    private String jobLocation;
-
-}
