--- Backend/src/main/java/com/GreenCommute/Backend/Dto/UserDto.java
@@ -1,16 +1,0 @@
-package com.GreenCommute.Backend.Dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserDto {
-
-    private int userId;
-
-    private String userName;
-
-}
