--- Backend/src/main/java/com/GreenCommute/Backend/JPARepository/UserJpa.java
@@ -1,9 +1,0 @@
-package com.GreenCommute.Backend.JPARepository;
-
-import com.GreenCommute.Backend.Entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserJpa extends JpaRepository<User, Integer> {
-}
