--- Backend/src/main/java/com/GreenCommute/Backend/Service/JobService.java
@@ -1,14 +1,0 @@
-package com.GreenCommute.Backend.Service;
-
-import com.GreenCommute.Backend.Entity.Jobs;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface JobService {
-    Optional<Jobs> getJobById(int id);
-
-    List<Jobs> getAllJobs();
-
-    List<Jobs> getJobsSearchByLocation(String location);
-}
