--- Backend/src/main/java/com/GreenCommute/Backend/Service/UserService.java
@@ -1,9 +1,0 @@
-package com.GreenCommute.Backend.Service;
-
-import com.GreenCommute.Backend.Entity.User;
-
-import java.util.Optional;
-
-public interface UserService {
-    Optional<User> getUserById(int userId);
-}
