+++ Backend/src/main/java/com/greencommute/backend/dto/ResponseDto.java
@@ -1,0 +1,18 @@
+package com.greencommute.backend.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ResponseDto {
+
+    private int jobId;
+
+    private int userId;
+
+    private String message;
+
+}
