+++ Backend/src/main/java/com/greencommute/backend/dto/UserDto.java
@@ -1,0 +1,16 @@
+package com.greencommute.backend.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserDto {
+
+    private int userId;
+
+    private String userName;
+
+}
