+++ Backend/src/main/java/com/greencommute/backend/exception/DataNotFoundException.java
@@ -1,0 +1,8 @@
+package com.greencommute.backend.exception;
+
+public class DataNotFoundException extends RuntimeException {
+    public DataNotFoundException(String message) {
+        super(message);
+    }
+
+}
