+++ Backend/src/main/java/com/greencommute/backend/service/UserService.java
@@ -1,0 +1,9 @@
+package com.greencommute.backend.service;
+
+import com.greencommute.backend.entity.User;
+
+import java.util.Optional;
+
+public interface UserService {
+    Optional<User> getUserById(int userId);
+}
