+++ Backend/src/main/java/com/greencommute/backend/service/UserServiceImpl.java
@@ -1,0 +1,20 @@
+package com.greencommute.backend.service;
+
+import com.greencommute.backend.entity.User;
+import com.greencommute.backend.repository.UserJpa;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+@Service
+public class UserServiceImpl implements UserService {
+
+    @Autowired
+    public UserJpa userJpa;
+
+    @Override
+    public Optional<User> getUserById(int userId) {
+        return userJpa.findById(userId);
+    }
+}
