--- Backend/src/test/java/com/GreenCommute/Backend/BackendApplicationTests.java
@@ -1,13 +1,0 @@
-package com.GreenCommute.Backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
