+++ Backend/src/test/java/com/greencommute/backend/BackendApplicationTests.java
@@ -1,0 +1,13 @@
+package com.greencommute.backend;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackendApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
